git-svn-id: http://mewsplayground-svn.cvsdude.com/seaweeds/trunk/bsponmpi@238 09f15e90-c184-4a26-aad6-528d2e8cadf7

+++ src/bsp_cpp/bsp_commandline.cpp
@@ -38,7 +38,7 @@ namespace bsp {
 				args.resize(argc);
 			}
 
-			for (int a = 0; a < argc; ++a) {
+			for (size_t a = 0; a < argc; ++a) {
 				bsp_broadcast(0, args[a]);
 			}
 		}
